refactor: 應使用 pips 命名資料夾
